+++ src/grongier/pex/__init__.py
@@ -7,6 +7,7 @@ from iop._inbound_adapter import _InboundAdapter
 from iop._outbound_adapter import _OutboundAdapter
 from iop._message import _Message
 from iop._message import _PickleMessage
+from iop._persistent_message import Field, _PersistentMessage
 from iop._director import _Director
 from iop._utils import _Utils
 
@@ -21,4 +22,6 @@ class DuplexOperation(_PrivateSessionDuplex): pass
 class DuplexProcess(_PrivateSessionProcess): pass
 class Message(_Message): pass
 class PickleMessage(_PickleMessage): pass
+class PersistentMessage(_PersistentMessage):
+    _iop_persistent_message_abstract = True
 class Director(_Director): pass